refactor: move: plancalendar/studygroup 이동
각각의 기능별 패키지를 plancalendar 로 이동

[skip ci]
