+++ server/edusync/src/main/java/com/codestates/edusync/model/study/plancalendar/repository/CalendarStudygroupRepository.java
@@ -1,4 +1,4 @@
-package com.codestates.edusync.model.study.plancalendar.studygroup.repository;
+package com.codestates.edusync.model.study.plancalendar.repository;
 
 import com.codestates.edusync.model.study.plancalendar.entity.TimeSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
