+++ server/edusync/src/main/java/com/codestates/edusync/model/study/plancalendar/service/CalendarStudygroupService.java
@@ -1,12 +1,12 @@
-package com.codestates.edusync.model.study.plancalendar.studygroup.service;
+package com.codestates.edusync.model.study.plancalendar.service;
 
 import com.codestates.edusync.model.common.utils.VerifyStudygroupCalendarUtils;
 import com.codestates.edusync.model.common.utils.VerifyVerifyStudygroupUtils;
 import com.codestates.edusync.model.member.entity.Member;
 import com.codestates.edusync.model.study.studygroup.entity.Studygroup;
 import com.codestates.edusync.model.study.plancalendar.entity.TimeSchedule;
-import com.codestates.edusync.model.study.plancalendar.studygroup.repository.CalendarStudygroupRepository;
-import com.codestates.edusync.model.study.plancalendar.studygroup.utils.CalendarStudygroupManager;
+import com.codestates.edusync.model.study.plancalendar.repository.CalendarStudygroupRepository;
+import com.codestates.edusync.model.study.plancalendar.utils.CalendarStudygroupManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
